Move store to util
